# private/ApiGrua/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Grua Laravel",
        "description": "Api de Grua para proyecto DWEC. Vue + Laravel",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost/DWEC/private/ApiGrua/public/"
        }
    ],
    "paths": {
        "/api/logs": {
            "get": {
                "tags": [
                    "Logs"
                ],
                "summary": "Muestra todos los logs",
                "operationId": "95e3308fea806ffc31774c0be0c7d05f",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Log"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Logs"
                ],
                "summary": "Registra un nuevo log",
                "operationId": "2bdb62e9fb5a08a508a2acaa2a3765a8",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "usuario_id": {
                                                "type": "integer"
                                            },
                                            "descripcion": {
                                                "type": "string"
                                            },
                                            "accion": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "usuario_id": 1,
                                    "descripcion": "Login de usuario 1",
                                    "accion": "LOGIN"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "CREATED",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "usuario_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "descripcion": {
                                            "type": "string",
                                            "example": "User login attempt"
                                        },
                                        "accion": {
                                            "type": "string",
                                            "example": "LOGIN"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2023-02-23T00:09:16.000000Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2023-02-23T12:33:45.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Contenido no válido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The usuario_id field is required."
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "Error object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/logs/{id}": {
            "get": {
                "tags": [
                    "Logs"
                ],
                "summary": "Obtiene un log específico por ID",
                "description": "Retorna la información detallada de un log específico",
                "operationId": "4002fbd2bdfe1d18a4000fea389e0e5b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del log a consultar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Log encontrado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "usuario_id": {
                                            "type": "integer"
                                        },
                                        "descripcion": {
                                            "type": "string"
                                        },
                                        "accion": {
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Log no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Log no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/tarifas/{id}": {
            "get": {
                "tags": [
                    "Tarifas"
                ],
                "summary": "Obtiene una tarifa específica por ID",
                "description": "Retorna la información detallada de una tarifa específica",
                "operationId": "6e99b791d13b4d9f8c8da4452fe447b2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la tarifa a consultar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tarifa encontrada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "horas_gratis": {
                                            "type": "string"
                                        },
                                        "costo_por_hora": {
                                            "type": "string"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tarifa no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tarifa no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Log": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "usuario_id": {
                        "type": "integer"
                    },
                    "descripcion": {
                        "type": "string"
                    },
                    "accion": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}